+++ infer/crates/e2k-rs/src/layers.rs
@@ -1,7 +1,7 @@
 use itertools::Itertools;
 use ndarray::prelude::*;
 
-pub(crate) fn sigmoid_1d(x: Array1<f32>) -> Array1<f32> {
+pub(crate) fn sigmoid_1d(x: Array1<f64>) -> Array1<f64> {
     x.map(|x| 1.0 / (1.0 + (-x).exp()))
 }
 
@@ -59,19 +59,19 @@ macro_rules! split_ndarray_owned {
 
 #[derive(Debug)]
 pub(crate) struct Linear {
-    weight: ndarray::Array2<f32>,
-    bias: ndarray::Array1<f32>,
+    weight: ndarray::Array2<f64>,
+    bias: ndarray::Array1<f64>,
 }
 
 impl Linear {
-    pub fn new(weight: ndarray::Array2<f32>, bias: ndarray::Array1<f32>) -> Self {
+    pub fn new(weight: ndarray::Array2<f64>, bias: ndarray::Array1<f64>) -> Self {
         Self { weight, bias }
     }
-    pub fn forward_2d(&self, input: &ndarray::Array2<f32>) -> ndarray::Array2<f32> {
+    pub fn forward_2d(&self, input: &ndarray::Array2<f64>) -> ndarray::Array2<f64> {
         let output = input.dot(&self.weight.t());
         output + &self.bias
     }
-    pub fn forward_1d(&self, input: &ndarray::Array1<f32>) -> ndarray::Array1<f32> {
+    pub fn forward_1d(&self, input: &ndarray::Array1<f64>) -> ndarray::Array1<f64> {
         let output = input.dot(&self.weight.t());
         output + &self.bias
     }
@@ -79,14 +79,14 @@ impl Linear {
 
 #[derive(Debug)]
 pub(crate) struct Embedding {
-    weight: ndarray::Array2<f32>,
+    weight: ndarray::Array2<f64>,
 }
 
 impl Embedding {
-    pub fn new(weight: ndarray::Array2<f32>) -> Self {
+    pub fn new(weight: ndarray::Array2<f64>) -> Self {
         Self { weight }
     }
-    pub fn forward(&self, input: &ndarray::Array1<usize>) -> ndarray::Array2<f32> {
+    pub fn forward(&self, input: &ndarray::Array1<usize>) -> ndarray::Array2<f64> {
         ndarray::stack(
             ndarray::Axis(0),
             &input
@@ -106,15 +106,15 @@ pub(crate) struct Mha {
     v_proj: Linear,
     out_proj: Linear,
     n_heads: usize,
-    scale: f32,
+    scale: f64,
 }
 
 impl Mha {
     pub(crate) fn new(
-        in_proj_weight: ndarray::Array2<f32>,
-        in_proj_bias: ndarray::Array1<f32>,
-        out_proj_weight: ndarray::Array2<f32>,
-        out_proj_bias: ndarray::Array1<f32>,
+        in_proj_weight: ndarray::Array2<f64>,
+        in_proj_bias: ndarray::Array1<f64>,
+        out_proj_weight: ndarray::Array2<f64>,
+        out_proj_bias: ndarray::Array1<f64>,
         num_heads: usize,
     ) -> Self {
         let (q_w, k_w, v_w) = split_ndarray!(&in_proj_weight, 3, ndarray::Axis(0));
@@ -125,7 +125,7 @@ impl Mha {
         let v_proj = Linear::new(v_w.to_owned(), v_b.to_owned());
         let out_proj = Linear::new(out_proj_weight, out_proj_bias);
         let d_heads = dim / num_heads;
-        let scale = (d_heads as f32).sqrt();
+        let scale = (d_heads as f64).sqrt();
         Self {
             q_proj,
             k_proj,
@@ -138,10 +138,10 @@ impl Mha {
 
     pub(crate) fn forward(
         &self,
-        query: &ndarray::Array2<f32>,
-        key: &ndarray::Array2<f32>,
-        value: &ndarray::Array2<f32>,
-    ) -> ndarray::Array2<f32> {
+        query: &ndarray::Array2<f64>,
+        key: &ndarray::Array2<f64>,
+        value: &ndarray::Array2<f64>,
+    ) -> ndarray::Array2<f64> {
         let q = self.q_proj.forward_2d(query);
         let k = self.k_proj.forward_2d(key);
         let v = self.v_proj.forward_2d(value);
@@ -177,10 +177,10 @@ pub(crate) struct GruCell {
 
 impl GruCell {
     pub(crate) fn new(
-        weight_ih: ndarray::Array2<f32>,
-        weight_hh: ndarray::Array2<f32>,
-        bias_ih: ndarray::Array1<f32>,
-        bias_hh: ndarray::Array1<f32>,
+        weight_ih: ndarray::Array2<f64>,
+        weight_hh: ndarray::Array2<f64>,
+        bias_ih: ndarray::Array1<f64>,
+        bias_hh: ndarray::Array1<f64>,
     ) -> Self {
         let ih = Linear::new(weight_ih, bias_ih);
         let hh = Linear::new(weight_hh, bias_hh);
@@ -189,9 +189,9 @@ impl GruCell {
 
     pub(crate) fn forward(
         &self,
-        input: &ndarray::Array1<f32>,
-        hidden: &Option<ndarray::Array1<f32>>,
-    ) -> ndarray::Array1<f32> {
+        input: &ndarray::Array1<f64>,
+        hidden: &Option<ndarray::Array1<f64>>,
+    ) -> ndarray::Array1<f64> {
         let hidden = hidden.clone().unwrap_or_else(|| {
             ndarray::Array1::zeros((self.hh.weight.shape()[self.hh.weight.ndim() - 1],))
         });
@@ -232,9 +232,9 @@ impl Gru {
 
     pub(crate) fn forward(
         &self,
-        input: &ndarray::Array2<f32>,
-        hidden: Option<ndarray::Array1<f32>>,
-    ) -> (ndarray::Array2<f32>, ndarray::Array1<f32>) {
+        input: &ndarray::Array2<f64>,
+        hidden: Option<ndarray::Array1<f64>>,
+    ) -> (ndarray::Array2<f64>, ndarray::Array1<f64>) {
         let mut hidden = hidden;
         let input = if self.reverse {
             input.slice(s![..; -1, ..]).to_owned()
